+++ frontend/package.json
@@ -19,7 +19,8 @@
     "bootstrap": "^5.3.3",
     "next": "14.1.0",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "react-hook-form": "^7.53.2"
   },
   "devDependencies": {
     "@types/axios": "^0.14.0",
